Move additional gameplay related hooks to multiplayer feature folder
